fix: deploy errors
