+++ src/components/custom/navbar/lib/index.ts
@@ -1,2 +1,2 @@
-export { defaultConfig } from './navbar-config.constant';
+export { defaultConfig, NAV_ITEMS } from './navbar-config.constant';
 export { TranslationKey } from './translation-keys.enum';
